+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/core/Wei.java
@@ -18,7 +18,7 @@ import org.hyperledger.besu.plugin.data.Quantity;
 
 import java.math.BigInteger;
 
-import org.apache.tuweni.bytes.Bytes32;
+import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.units.bigints.BaseUInt256Value;
 import org.apache.tuweni.units.bigints.UInt256;
 
@@ -61,7 +61,7 @@ public final class Wei extends BaseUInt256Value<Wei> implements Quantity {
     return new Wei((BigInteger) value);
   }
 
-  public static Wei wrap(final Bytes32 value) {
+  public static Wei wrap(final Bytes value) {
     return new Wei(UInt256.fromBytes(value));
   }