+++ ethereum/eth/src/main/java/org/hyperledger/besu/ethereum/eth/messages/LimitedTransactionsMessages.java
@@ -45,17 +45,14 @@ public final class LimitedTransactionsMessages {
       final BytesValueRLPOutput encodedTransaction = new BytesValueRLPOutput();
       transaction.writeTo(encodedTransaction);
       Bytes encodedBytes = encodedTransaction.encoded();
-      // Break if individual transaction size exceeds limit
-      if (encodedBytes.size() > LIMIT && (messageSize != 0)) {
+      if (messageSize != 0 // always at least one message
+          && messageSize + encodedBytes.size() > LIMIT) {
         break;
       }
       message.writeRaw(encodedBytes);
       includedTransactions.add(transaction);
       // Check if last transaction to add to the message
       messageSize += encodedBytes.size();
-      if (messageSize > LIMIT) {
-        break;
-      }
     }
     message.endList();
     return new LimitedTransactionsMessages(